split vis reg from simpler codechecks, replace custom commit-deployment with external check
